third_party/WebKit/LayoutTests/fast/css/fontfaceset-add-remove-expected.txt

Tests that adding/removing a FontFace to/from FontFaceSet trigger font update.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testElement.offsetWidth is arialElement.offsetWidth
PASS addResult === document.fonts is true
PASS testElement.offsetWidth is ahemElement.offsetWidth
PASS testElement.offsetWidth is arialElement.offsetWidth
PASS testElement.offsetWidth is ahemElement.offsetWidth
PASS testElement.offsetWidth is arialElement.offsetWidth
PASS successfullyParsed is true

TEST COMPLETE

abcdefg 
abcdefg 
abcdefg
